supabase-fastapi/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
openai>=1.0,<2.0
supabase>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
passlib[bcrypt]>=1.7,<2.0
python-multipart>=0.0.9,<1.0
pydantic>=2.0,<3.0
